# openapi.yaml
openapi: 3.1.0
info:
  title: document-pipeline
  version: 1.0.0
  description: |
    Ingests documents through a configurable OCR → clarify → classify → embed pipeline,
    and stores the results in a vector database for RAG queries.

    Documents move through pipeline stages asynchronously via jobs. Each job represents
    execution of a single stage. The frontend polls for state changes and streams live
    LLM output via Server-Sent Events.
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - authentik: []
...
paths:
...
  /api/v1/documents:
    get:
      tags: [Documents]
      summary: List documents
      description: Returns a paginated list of document summaries (no artifacts or context).
      operationId: listDocuments
      parameters:
        - $ref: '#/components/parameters/page_token'
        - $ref: '#/components/parameters/page_size'
        - $ref: '#/components/parameters/sort_order'
        - $ref: '#/components/parameters/stages_filter'
        - $ref: '#/components/parameters/statuses_filter'
      responses:
        '200':
          description: Paginated list of document summaries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDocuments'
    post:
      tags: [Documents]
      summary: Upload a document
      description: Upload a .txt, .md, or image file to create a new document. Text files skip the OCR stage.
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                  nullable: true
                additional_context:
                  type: string
                  nullable: true
                linked_contexts:
                  type: array
                  items:
                    type: string
                    format: uuid
                  nullable: true
                series:
                  type: string
                  nullable: true
                  description: Series name for grouping related documents into a shared embedding corpus.
      responses:
        '201':
          description: Document created. Returns initial job detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDetail'
        '400':
          description: Unsupported file type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate file — document already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/documents/{doc_id}:
    parameters:
      - $ref: '#/components/parameters/doc_id'
    get:
      tags: [Documents]
      summary: Get document detail
      description: Returns full document data — title, context, and artifacts. Does not include job/pipeline state.
      operationId: getDocument
      responses:
        '200':
          description: Document detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Documents]
      summary: Update document
      description: Updates one or more mutable document fields. Only fields present in the request body are changed.
      operationId: patchDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchDocumentBody'
      responses:
        '200':
          description: Updated document detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Documents]
      summary: Delete document
      description: Permanently deletes the document, all artifacts, and all jobs.
      operationId: deleteDocument
      responses:
        '200':
          description: Document deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/documents/{doc_id}/artifacts/{artifact_id}:
    parameters:
      - $ref: '#/components/parameters/doc_id'
      - $ref: '#/components/parameters/artifact_id'
    get:
      tags: [Documents]
      summary: Get artifact content
      description: Returns the raw bytes of an artifact with the correct Content-Type header. Use this to display source images, OCR outputs, or any other file produced during pipeline execution.
      operationId: getArtifact
      responses:
        '200':
          description: Artifact bytes.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Jobs ─────────────────────────────────────────────────────────────────────
...
components:
...
  parameters:
    doc_id:
      name: doc_id
      in: path
      required: true
      description: UUID of the document.
      schema:
        type: string
        format: uuid
...
    artifact_id:
      name: artifact_id
      in: path
      required: true
      description: UUID of the artifact.
      schema:
        type: string
        format: uuid
...
    page_token:
      name: page_token
      in: query
      required: false
      description: Opaque cursor token returned by the previous page response. Omit for the first page.
      schema:
        type: string
    page_size:
      name: page_size
      in: query
      required: false
      description: Maximum number of results to return (default 50, max 1000).
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 1000
    stages_filter:
      name: stages
      in: query
      required: false
      description: Comma-separated list of pipeline stage names to filter by.
      schema:
        type: string
        example: ocr,clarify
    statuses_filter:
      name: statuses
      in: query
      required: false
      description: Comma-separated list of job statuses to filter by.
      schema:
        type: string
        example: pending,waiting
    sort_order:
      name: sort
      in: query
      required: false
      description: Sort order.
      schema:
        type: string
        default: pipeline
        enum: [pipeline, title_asc, title_desc, created_asc, created_desc]

  # ── Reusable responses ────────────────────────────────────────────────────────
  responses:
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ── Schemas ───────────────────────────────────────────────────────────────────
  schemas:
...
    JobStatus:
      type: string
      description: Execution status of a job.
      enum: [pending, running, waiting, error, done]
    ErrorResponse:
      type: object
      description: Generic error body.
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
    OkResponse:
      type: object
      description: Simple success acknowledgement.
      required: [ok]
      properties:
        ok:
          type: boolean
          description: Always true on success.

    # — Pipelines ——————————————————————————————————————————————————————————————
...
    Artifact:
      type: object
      description: |
        A file artifact associated with a document. Source images uploaded by the user
        are stored as artifacts with `created_job_id: null`. Pipeline stages (e.g. OCR)
        may produce additional artifacts such as extracted text files.
      required: [id, filename, content_type, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          description: Unique artifact identifier.
        filename:
          type: string
          description: Original filename of the artifact.
        content_type:
          type: string
          description: MIME type (e.g. `image/png`, `text/plain`).
        created_job_id:
          type: string
          format: uuid
          nullable: true
          description: UUID of the job that created this artifact. Null for source artifacts created during ingest.
        created_at:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp.
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 last-updated timestamp.
    DocumentSummary:
      type: object
      description: Lightweight document representation used in list views.
      required: [id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          description: Unique document identifier.
        title:
          type: string
          nullable: true
          description: Human-readable title.
        current_job_id:
          type: string
          format: uuid
          nullable: true
          description: UUID of the most recently active job for this document.
        series:
          type: string
          nullable: true
          description: Series name grouping this document with related documents for corpus embedding.
        created_at:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp.
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 last-updated timestamp.
    DocumentDetail:
      allOf:
        - $ref: '#/components/schemas/DocumentSummary'
        - type: object
          description: Full document data including context and artifacts.
          properties:
            additional_context:
              type: string
              description: Free-text context attached by the user to guide LLM stages.
            linked_contexts:
              type: array
              items:
                type: string
                format: uuid
              description: UUIDs of saved context entries linked to this document.
            artifacts:
              type: array
              items:
                $ref: '#/components/schemas/Artifact'
              description: Files associated with this document (e.g. source image).
    PaginatedDocuments:
      type: object
      description: Paginated list of document summaries.
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSummary'
        next_page_token:
          type: string
          nullable: true
          description: Opaque cursor; pass as `page_token` on next request. Null when no more pages.
    PatchDocumentBody:
      type: object
      description: Fields to update on a document. Only present fields are applied.
      properties:
        title:
          type: string
          nullable: true
          description: Updated document title. Null clears the title.
        additional_context:
          type: string
          nullable: true
          description: Free-text context to attach to the document. Null clears it.
        linked_contexts:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
          description: UUIDs of context entries to link. Null clears all linked contexts.
        series:
          type: string
          nullable: true
          description: Series name. Null clears the series assignment.

    # — Jobs ———————————————————————————————————————————————————————————————————
    RunIOField:
      type: object
      description: A single input or output field for an LLM run.
      required: [field, text]
      properties:
        field:
          type: string
          nullable: true
          description: Field name (e.g. `ocr_raw`, `clarified_text`).
        text:
          type: string
          description: Field value.
    RunQuestion:
      type: object
      description: A clarification question raised by the LLM, with optional answer.
      required: [segment, question]
      properties:
        segment:
          type: string
          description: The text segment that prompted the question.
        question:
          type: string
          description: The clarification question to present to the user.
        answer:
          type: string
          nullable: true
          description: The user's answer. Null until answered.
    RunSuggestions:
      type: object
      description: LLM-proposed updates to document and linked context.
      properties:
        additional_context:
          type: string
          description: Suggested update to the document's additional context.
        linked_context:
          type: string
          description: Suggested update to the linked context entry text.
        linked_context_id:
          type: string
          format: uuid
          nullable: true
          description: UUID of the linked context entry this suggestion applies to.
    Run:
      type: object
      description: |
        A single LLM round trip within a job execution. Jobs accumulate runs over their
        lifetime — each execution (initial run, retry, or re-run after clarification)
        appends a new run to the list.
      required: [id, inputs, outputs, confidence, questions, suggestions, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          description: Unique run identifier.
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/RunIOField'
          description: Fields fed into the LLM.
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/RunIOField'
          description: Fields produced by the LLM.
        confidence:
          type: string
          enum: [high, medium, low]
          description: LLM-reported confidence in the output.
        questions:
          type: array
          items:
            $ref: '#/components/schemas/RunQuestion'
          description: Clarification questions raised by the LLM.
        suggestions:
          $ref: '#/components/schemas/RunSuggestions'
        created_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp when this run started.
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp when this run was last updated.
    EmbedOptions:
      type: object
      description: Options specific to the embed stage.
      properties:
        embed_image:
          type: boolean
          default: false
          description: Whether to include a visual image vector in addition to the text vector.
    JobOptions:
      type: object
      description: Per-job configuration options.
      properties:
        require_context:
          type: boolean
          default: false
          description: Whether this stage requires context before it can run.
        embed:
          $ref: '#/components/schemas/EmbedOptions'
    JobSummary:
      type: object
      description: Lightweight job representation used in list views.
      required: [id, document_id, stage, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          description: Unique job identifier.
        document_id:
          type: string
          format: uuid
          description: UUID of the document this job belongs to.
        title:
          type: string
          nullable: true
          description: Denormalized document title.
        stage:
          type: string
          description: Pipeline stage this job executes (e.g. `ocr`, `clarify`).
        status:
          $ref: '#/components/schemas/JobStatus'
        created_at:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp.
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 last-updated timestamp.
    JobDetail:
      allOf:
        - $ref: '#/components/schemas/JobSummary'
        - type: object
          description: Full job data including options and run history.
          properties:
            options:
              $ref: '#/components/schemas/JobOptions'
            runs:
              type: array
              items:
                $ref: '#/components/schemas/Run'
              description: Ordered list of LLM round trips for this job. Appended on each execution.
...
  securitySchemes:
    authentik:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.jasonfagerberg.duckdns.org/application/o/authorize/
          tokenUrl: https://auth.jasonfagerberg.duckdns.org/application/o/token/
          refreshUrl: https://auth.jasonfagerberg.duckdns.org/application/o/token/
          scopes:
            openid: OpenID Connect identity token
            profile: User profile (name, username)
            email: User email address